Migrated SharedResources to unified api
Pasted nint to int. Changed
NavigationController.DismissModalViewControllerAnimated to
NavigationController.DismissModalViewcontroller. Changed
NavigationController.PopViewControllerAnimated to
NavigationController.PopViewController. Changed
CGContext.SetRGBFillColor to CGContext.SetFillColor.

+++ SharedResources/SharedResources/Code/NavItemTableSource.cs
@@ -1,7 +1,7 @@
 using System;
 using System.Collections.Generic;
-using MonoTouch.UIKit;
-using MonoTouch.Foundation;
+using UIKit;
+using Foundation;
 using System.Reflection;
 
 namespace Example_SharedResources.Code
@@ -24,7 +24,7 @@ namespace Example_SharedResources.Code
 		/// <summary>
 		/// Called by the TableView to determine how many sections(groups) there are.
 		/// </summary>
-		public override int NumberOfSections (UITableView tableView)
+		public override nint NumberOfSections (UITableView tableView)
 		{
 			return navItems.Count;
 		}
@@ -32,25 +32,28 @@ namespace Example_SharedResources.Code
 		/// <summary>
 		/// Called by the TableView to determine how many cells to create for that particular section.
 		/// </summary>
-		public override int RowsInSection (UITableView tableview, int section)
+		public override nint RowsInSection (UITableView tableview, nint section)
 		{
-			return navItems[section].Items.Count;
+			//HACK: Parse section to int
+			return navItems[(int)section].Items.Count;
 		}
 
 		/// <summary>
 		/// Called by the TableView to retrieve the header text for the particular section(group)
 		/// </summary>
-		public override string TitleForHeader (UITableView tableView, int section)
+		public override string TitleForHeader (UITableView tableView, nint section)
 		{
-			return navItems[section].Name;
+			//HACK: Parse section to int
+			return navItems[(int)section].Name;
 		}
 
 		/// <summary>
 		/// Called by the TableView to retrieve the footer text for the particular section(group)
 		/// </summary>
-		public override string TitleForFooter (UITableView tableView, int section)
+		public override string TitleForFooter (UITableView tableView, nint section)
 		{
-			return navItems[section].Footer;
+			//HACK: Parse section to int
+			return navItems[(int)section].Footer;
 		}
 
 		/// <summary>
